requirements.txt
## general
semver
python-dotenv
logfmter

## web
fastapi[standard]
redis
minio

## steam
steam[client]
gevent

## job
celery
celery-singleton
flower